Commit #5
-Completed Customers and Customer Policy
- Added Invoices and Invoice Policy

+++ app/Http/Controllers/CustomerController.php
@@ -2,8 +2,10 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\customer;
+use App\Models\Customer;
+use App\Models\Company;
 use Illuminate\Http\Request;
+use Auth;
 
 class CustomerController extends Controller
 {
@@ -14,8 +16,26 @@ class CustomerController extends Controller
      */
     public function index()
     {
-        $customers = customer::all();
-        return view('customers.index', compact('customers'));
+        $this->authorize('viewAny',Customer::class);
+        $user = Auth::user();
+        $user_id = $user->id;
+        $role = Auth::user()->role;
+        //IF SUPER ADMIN GET ALL COMPANIES
+        if($user_id == 1){
+            $customers = Customer::all();
+        }else{
+            //IF WEBMASTER GET PRODUCTS OWNED BY COMPMNY THAT WEBMASTER BELONGS TO
+            if($role->id == $user->isWebMaster()){
+                $companies = Company::where('user_id', $user_id)->get('id');
+                $customers = Customer::whereIn('company_id',$companies)->get();
+            }
+            //IF OFFICER GET PRODUCTS OWNED BY COMPMNY THAT OFFICER ASSIGNED TO
+            elseif($role->id == $user->isOfficer()){
+                $companies = $user->companies->pluck('id');
+                $customers = Customer::whereIn('company_id',$companies)->get();
+            }
+        }
+        return view('customers.index',compact('customers'));
     }
 
     /**
@@ -25,7 +45,22 @@ class CustomerController extends Controller
      */
     public function create()
     {
-        return view('customers.create');
+        $this->authorize('create',Customer::class);
+        $user = Auth::user();
+        $user_id = $user->id;
+        $role = Auth::user()->role;
+        
+        if($user_id == 1){
+            $companies = Company::all();
+        }else{
+            if($role->id == $user->isWebMaster()){
+                $companies = Company::where('user_id', $user_id)->get();
+            }
+            if($role->id == $user->isOfficer()){
+                $companies = $user->companies;
+            }
+        }
+        return view('customers.create',compact('companies'));
     }
 
     /**
@@ -36,6 +71,7 @@ class CustomerController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create',Customer::class);
         $rules = [
             'first_name' => 'required|max:255',
             'last_name' => 'required|max:255',
@@ -44,6 +80,7 @@ class CustomerController extends Controller
             'phone' => 'required|max:255',
             'mobile' => 'max:255',
             'address' => 'required|max:255',
+            'company_id' => 'required|not in:Open this select menu',
         ];
 
         $this->validate($request, $rules);
@@ -71,7 +108,22 @@ class CustomerController extends Controller
      */
     public function edit(customer $customer)
     {
-        return view('customers.edit',compact('customer'));
+        $this->authorize('update',$customer);
+        $user = Auth::user();
+        $user_id = $user->id;
+        $role = Auth::user()->role;
+
+        if($user_id == 1){
+            $companies = Company::all();
+        }else{
+            if($role->id == $user->isWebMaster()){
+                $companies = Company::where('user_id', $user_id)->get();
+            }
+            if($role->id == $user->isOfficer()){
+                $companies = $user->companies;
+            }
+        }
+        return view('customers.edit',compact('customer','companies'));
     }
 
     /**
@@ -83,6 +135,7 @@ class CustomerController extends Controller
      */
     public function update(Request $request, customer $customer)
     {
+        $this->authorize('update',$customer);
         $rules = [
             'first_name' => 'required|max:255',
             'last_name' => 'required|max:255',
@@ -106,16 +159,16 @@ class CustomerController extends Controller
      * @param  \App\Models\customer  $customer
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Customer $customer)
     {
-        $customer = customer::findOrFail($id);
+        $this->authorize('delete',$customer);
         $customer->delete();
         return redirect()->route('customers.index');
     }
 
-    public function delete($id)
+    public function delete(Customer $customer)
     {
-        $customer = customer::findOrFail($id);
+        $this->authorize('delete',$customer);
         return view('customers.delete', ['customer'=>$customer]);
     }
 }
